Split tests into two files (#1116)
The `test_preprocess.py` file has gotten too long (over 1000 lines), and
CodeFactor doesn't like it. We split it into two files, one for the
measurement transforms, and one for the operator decomposition
transforms.
